updates for actor deferred proxy
This commit essentially breaks the old model of Worker::Actor and
Worker::Actor::Impl so that we don't need to hold an extra strong
reference to Worker::Actor.

Worker::Actor is more so a Worker::TimedActorRef now, and
Worker::Actor::Impl is more accurately described as Worker::Actor.
Worker::Actor::Impl being forward declared is kind of breaking things
at the moment, so the next commit will change its name and bring it up a
level (Worker::ActorImpl).